--- grafana/config.monitoring
@@ -1,3 +1,0 @@
-GF_SECURITY_ADMIN_PASSWORD=admin
-GF_USERS_ALLOW_SIGN_UP=false
-GF_INSTALL_PLUGINS=grafana-piechart-panel
