--- prometheus/prometheus.yml
@@ -1,10 +1,0 @@
-global:
-  scrape_interval:     15s
-  evaluation_interval: 15s
-
-scrape_configs:
-  - job_name: 'traefik'
-    scrape_interval: 5s
-
-    static_configs:
-         - targets: ['traefik:8080']
